+++ src/collective/volto/formsupport/tests/test_captcha.py
@@ -24,7 +24,6 @@ import unittest
 
 
 class TestCaptcha(unittest.TestCase):
-
     layer = VOLTO_FORMSUPPORT_API_FUNCTIONAL_TESTING
 
     def setUp(self):
@@ -85,7 +84,7 @@ class TestCaptcha(unittest.TestCase):
                 "@type": "form",
                 "default_subject": "block subject",
                 "default_from": "john@doe.com",
-                "send": True,
+                "send": ["recipient"],
                 "subblocks": [
                     {
                         "field_id": "contact",
@@ -126,7 +125,7 @@ class TestCaptcha(unittest.TestCase):
                 "@type": "form",
                 "default_subject": "block subject",
                 "default_from": "john@doe.com",
-                "send": True,
+                "send": ["recipient"],
                 "subblocks": [
                     {
                         "field_id": "contact",
@@ -203,7 +202,7 @@ class TestCaptcha(unittest.TestCase):
                 "@type": "form",
                 "default_subject": "block subject",
                 "default_from": "john@doe.com",
-                "send": True,
+                "send": ["recipient"],
                 "subblocks": [
                     {
                         "field_id": "contact",
@@ -324,7 +323,7 @@ class TestCaptcha(unittest.TestCase):
                 "@type": "form",
                 "default_subject": "block subject",
                 "default_from": "john@doe.com",
-                "send": True,
+                "send": ["recipient"],
                 "subblocks": [
                     {
                         "field_id": "contact",
@@ -371,7 +370,7 @@ class TestCaptcha(unittest.TestCase):
                 "@type": "form",
                 "default_subject": "block subject",
                 "default_from": "john@doe.com",
-                "send": True,
+                "send": ["recipient"],
                 "subblocks": [
                     {
                         "field_id": "contact",