+++ src/collective/volto/formsupport/tests/test_event.py
@@ -24,7 +24,6 @@ def event_handler(event):
 
 
 class TestEvent(unittest.TestCase):
-
     layer = VOLTO_FORMSUPPORT_API_FUNCTIONAL_TESTING
 
     def setUp(self):
@@ -91,14 +90,13 @@ class TestEvent(unittest.TestCase):
     def test_trigger_event(
         self,
     ):
-
         self.document.blocks = {
             "text-id": {"@type": "text"},
             "form-id": {
                 "@type": "form",
                 "default_subject": "block subject",
                 "default_from": "john@doe.com",
-                "send": True,
+                "send": ["recipient"],
                 "subblocks": [
                     {
                         "field_id": "contact",