+++ src/collective/volto/formsupport/tests/test_honeypot.py
@@ -18,7 +18,6 @@ import json
 
 
 class TestHoneypot(unittest.TestCase):
-
     layer = VOLTO_FORMSUPPORT_API_FUNCTIONAL_TESTING
 
     def setUp(self):
@@ -72,14 +71,13 @@ class TestHoneypot(unittest.TestCase):
         return response
 
     def test_honeypot_installed_but_field_not_in_form(self):
-
         self.document.blocks = {
             "text-id": {"@type": "text"},
             "form-id": {
                 "@type": "form",
                 "default_subject": "block subject",
                 "default_from": "john@doe.com",
-                "send": True,
+                "send": ["recipient"],
                 "subblocks": [
                     {
                         "field_id": "contact",
@@ -107,14 +105,13 @@ class TestHoneypot(unittest.TestCase):
         )
 
     def test_honeypot_field_in_form_empty_pass_validation(self):
-
         self.document.blocks = {
             "text-id": {"@type": "text"},
             "form-id": {
                 "@type": "form",
                 "default_subject": "block subject",
                 "default_from": "john@doe.com",
-                "send": True,
+                "send": ["recipient"],
                 "subblocks": [
                     {
                         "field_id": "contact",
@@ -138,14 +135,13 @@ class TestHoneypot(unittest.TestCase):
         self.assertEqual(response.status_code, 204)
 
     def test_honeypot_field_in_form_compiled_fail_validation(self):
-
         self.document.blocks = {
             "text-id": {"@type": "text"},
             "form-id": {
                 "@type": "form",
                 "default_subject": "block subject",
                 "default_from": "john@doe.com",
-                "send": True,
+                "send": ["recipient"],
                 "subblocks": [
                     {
                         "field_id": "contact",
@@ -183,7 +179,7 @@ class TestHoneypot(unittest.TestCase):
                 "@type": "form",
                 "default_subject": "block subject",
                 "default_from": "john@doe.com",
-                "send": True,
+                "send": ["recipient"],
                 "subblocks": [
                     {
                         "field_id": "contact",
@@ -222,7 +218,7 @@ class TestHoneypot(unittest.TestCase):
                 "@type": "form",
                 "default_subject": "block subject",
                 "default_from": "john@doe.com",
-                "send": True,
+                "send": ["recipient"],
                 "subblocks": [
                     {
                         "field_id": "contact",
@@ -264,7 +260,7 @@ class TestHoneypot(unittest.TestCase):
                 "@type": "form",
                 "default_subject": "block subject",
                 "default_from": "john@doe.com",
-                "send": True,
+                "send": ["recipient"],
                 "subblocks": [
                     {
                         "field_id": "contact",