+++ CHANGES.rst
@@ -10,12 +10,14 @@ Changelog
   [JeffersonBledsoe]
 - Add Spanish translation.
   [macagua]
-
+- Add German translation.
+  [mbarde]
 - Allow forwarding request headers in the sent emails #27
   [JeffersonBledsoe]
 - Added support for sending emails as a table #31
   [JeffersonBledsoe]
-
+- Add validation for email fields.
+  [cekk]
 
 2.7.0 (2023-04-03)
 ------------------